refactor(presets): move most ai related stuff to presets
