+++ packages/blocks/src/_common/components/index.ts
@@ -1,4 +1,3 @@
-export * from './ai-item/index.js';
 export * from './block-selection.js';
 export * from './drag-indicator.js';
 export * from './file-drop-manager.js';
