+++ packages/blocks/src/_common/icons/index.ts
@@ -1,4 +1,3 @@
-export * from './ai.js';
 export * from './database.js';
 export * from './edgeless.js';
 export * from './file-icons.js';
