+++ packages/blocks/src/root-block/widgets/index.ts
@@ -1,15 +1,6 @@
-export {
-  AFFINE_AI_PANEL_WIDGET,
-  AffineAIPanelWidget,
-  type AffineAIPanelWidgetConfig,
-} from './ai-panel/ai-panel.js';
 export { BlockHubWidget } from './block-hub/block-hub.js';
 export { AffineDocRemoteSelectionWidget } from './doc-remote-selection/doc-remote-selection.js';
 export { AffineDragHandleWidget } from './drag-handle/drag-handle.js';
-export {
-  AFFINE_EDGELESS_COPILOT_WIDGET,
-  EdgelessCopilotWidget,
-} from './edgeless-copilot/index.js';
 export { EdgelessRemoteSelectionWidget } from './edgeless-remote-selection/index.js';
 export { AffineEdgelessZoomToolbarWidget } from './edgeless-zoom-toolbar/index.js';
 export { toolbarDefaultConfig } from './format-bar/config.js';
