+++ packages/presets/src/ai/_common/icons.ts
@@ -2,39 +2,6 @@ import '@dotlottie/player-component';
 
 import { html } from 'lit';
 
-export const AIStarIcon = html`<svg
-  width="20"
-  height="20"
-  viewBox="0 0 20 20"
-  fill="none"
-  xmlns="http://www.w3.org/2000/svg"
->
-  <g clip-path="url(#clip0_7444_26654)">
-    <path
-      d="M9.40033 4.57585C9.36627 4.27519 9.11207 4.04792 8.80948 4.04761C8.50689 4.0473 8.25223 4.27405 8.21756 4.57465C7.99211 6.52928 7.40771 7.90373 6.46492 8.84651C5.52213 9.7893 4.14769 10.3737 2.19305 10.5991C1.89246 10.6338 1.66571 10.8885 1.66602 11.1911C1.66633 11.4937 1.8936 11.7479 2.19426 11.7819C4.11659 11.9997 5.52025 12.5838 6.48542 13.5315C7.44696 14.4757 8.04356 15.8498 8.21597 17.7908C8.24327 18.0981 8.50096 18.3337 8.80954 18.3333C9.11813 18.333 9.37528 18.0969 9.40189 17.7894C9.56714 15.8804 10.1632 14.4774 11.1295 13.5111C12.0958 12.5448 13.4988 11.9487 15.4078 11.7835C15.7153 11.7569 15.9514 11.4997 15.9517 11.1911C15.9521 10.8825 15.7165 10.6249 15.4092 10.5976C13.4683 10.4252 12.0941 9.82855 11.15 8.86701C10.2022 7.90184 9.61807 6.49818 9.40033 4.57585Z"
-      fill="#1E96EB"
-    />
-    <path
-      d="M15.7849 1.87208C15.7717 1.75516 15.6728 1.66678 15.5551 1.66666C15.4375 1.66654 15.3384 1.75472 15.3249 1.87162C15.2373 2.63175 15.01 3.16626 14.6434 3.5329C14.2767 3.89954 13.7422 4.1268 12.9821 4.21448C12.8652 4.22796 12.777 4.327 12.7771 4.44467C12.7772 4.56234 12.8656 4.6612 12.9826 4.67445C13.7301 4.75912 14.276 4.9863 14.6513 5.35485C15.0253 5.72203 15.2573 6.25642 15.3243 7.01121C15.335 7.13075 15.4352 7.22235 15.5552 7.22221C15.6752 7.22208 15.7752 7.13025 15.7855 7.01069C15.8498 6.26831 16.0816 5.72268 16.4574 5.3469C16.8332 4.97111 17.3788 4.73932 18.1212 4.67505C18.2407 4.6647 18.3325 4.5647 18.3327 4.44469C18.3328 4.32469 18.2412 4.22448 18.1217 4.21386C17.3669 4.14681 16.8325 3.9148 16.4653 3.54087C16.0968 3.16552 15.8696 2.61966 15.7849 1.87208Z"
-      fill="#1E96EB"
-    />
-  </g>
-  <defs>
-    <clipPath id="clip0_7444_26654">
-      <rect width="20" height="20" fill="white" />
-    </clipPath>
-  </defs>
-</svg>`;
-
-export const AIStarIconWithAnimation = html`<dotlottie-player
-  src='{"v":"5.12.1","fr":120,"ip":0,"op":239,"w":240,"h":240,"nm":"ai","ddd":0,"assets":[],"layers":[{"ddd":0,"ind":1,"ty":4,"nm":"big 2","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":38,"s":[80]},{"t":76,"s":[100]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[107.859,132.143,0],"ix":2,"l":2},"a":{"a":0,"k":[8.572,11.429,0],"ix":1,"l":2},"s":{"a":1,"k":[{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":38,"s":[0,0,100]},{"t":76,"s":[1005,1005,100]}],"ix":6,"l":2}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0.261,2.307],[0.363,0],[0.041,-0.361],[1.131,-1.132],[2.346,-0.271],[-0.001,-0.363],[-0.361,-0.041],[-1.158,-1.137],[-0.207,-2.329],[-0.37,0],[-0.032,0.369],[-1.16,1.16],[-2.291,0.198],[-0.001,0.371],[0.368,0.033],[1.133,1.154]],"o":[[-0.041,-0.361],[-0.363,0],[-0.271,2.345],[-1.132,1.131],[-0.36,0.042],[0,0.363],[2.307,0.262],[1.154,1.133],[0.033,0.369],[0.37,0],[0.198,-2.29],[1.16,-1.159],[0.369,-0.032],[0,-0.37],[-2.329,-0.207],[-1.138,-1.158]],"v":[[0.71,-7.937],[0.001,-8.571],[-0.709,-7.938],[-2.812,-2.812],[-7.939,-0.709],[-8.571,0.001],[-7.937,0.71],[-2.788,2.81],[-0.711,7.921],[0.001,8.572],[0.712,7.919],[2.785,2.785],[7.919,0.712],[8.572,0.001],[7.921,-0.711],[2.81,-2.788]],"c":true},"ix":2},"nm":"路径 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.117647058824,0.588235294118,0.921568627451,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"填充 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[8.572,11.429],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"变换"}],"nm":"组 1","np":4,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":8400,"st":0,"ct":1,"bm":0},{"ddd":0,"ind":2,"ty":4,"nm":"big","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":0,"s":[100]},{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":15,"s":[80]},{"t":38,"s":[0]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[107.859,132.143,0],"ix":2,"l":2},"a":{"a":0,"k":[8.572,11.429,0],"ix":1,"l":2},"s":{"a":1,"k":[{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":0,"s":[1005,1005,100]},{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":15,"s":[1035,1035,100]},{"t":38,"s":[0,0,100]}],"ix":6,"l":2}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0.261,2.307],[0.363,0],[0.041,-0.361],[1.131,-1.132],[2.346,-0.271],[-0.001,-0.363],[-0.361,-0.041],[-1.158,-1.137],[-0.207,-2.329],[-0.37,0],[-0.032,0.369],[-1.16,1.16],[-2.291,0.198],[-0.001,0.371],[0.368,0.033],[1.133,1.154]],"o":[[-0.041,-0.361],[-0.363,0],[-0.271,2.345],[-1.132,1.131],[-0.36,0.042],[0,0.363],[2.307,0.262],[1.154,1.133],[0.033,0.369],[0.37,0],[0.198,-2.29],[1.16,-1.159],[0.369,-0.032],[0,-0.37],[-2.329,-0.207],[-1.138,-1.158]],"v":[[0.71,-7.937],[0.001,-8.571],[-0.709,-7.938],[-2.812,-2.812],[-7.939,-0.709],[-8.571,0.001],[-7.937,0.71],[-2.788,2.81],[-0.711,7.921],[0.001,8.572],[0.712,7.919],[2.785,2.785],[7.919,0.712],[8.572,0.001],[7.921,-0.711],[2.81,-2.788]],"c":true},"ix":2},"nm":"路径 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.117647058824,0.588235294118,0.921568627451,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"填充 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[8.572,11.429],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"变换"}],"nm":"组 1","np":4,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":8400,"st":0,"ct":1,"bm":0},{"ddd":0,"ind":3,"ty":4,"nm":"small 2","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":48.074,"s":[80]},{"t":85.07421875,"s":[100]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[183,56.999,0],"ix":2,"l":2},"a":{"a":0,"k":[16.667,3.333,0],"ix":1,"l":2},"s":{"a":1,"k":[{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":48,"s":[0,0,100]},{"t":85.07421875,"s":[1005,1005,100]}],"ix":6,"l":2}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0.102,0.897],[0.141,0.001],[0.016,-0.14],[0.44,-0.44],[0.912,-0.105],[0,-0.141],[-0.14,-0.016],[-0.451,-0.442],[-0.08,-0.906],[-0.144,0],[-0.013,0.144],[-0.451,0.451],[-0.89,0.077],[0,0.144],[0.143,0.013],[0.441,0.448]],"o":[[-0.016,-0.14],[-0.141,0],[-0.105,0.913],[-0.44,0.44],[-0.141,0.016],[0,0.141],[0.897,0.101],[0.448,0.441],[0.013,0.143],[0.144,0],[0.077,-0.891],[0.451,-0.451],[0.144,-0.013],[0,-0.144],[-0.906,-0.08],[-0.442,-0.451]],"v":[[0.276,-3.087],[0,-3.334],[-0.276,-3.088],[-1.094,-1.094],[-3.087,-0.276],[-3.333,0],[-3.087,0.276],[-1.084,1.092],[-0.277,3.08],[0,3.333],[0.277,3.079],[1.083,1.083],[3.079,0.277],[3.333,0],[3.08,-0.277],[1.092,-1.084]],"c":true},"ix":2},"nm":"路径 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.117647058824,0.588235294118,0.921568627451,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"填充 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[16.667,3.333],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"变换"}],"nm":"组 1","np":4,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":8400,"st":0,"ct":1,"bm":0},{"ddd":0,"ind":4,"ty":4,"nm":"small","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":10.074,"s":[100]},{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":25.074,"s":[80]},{"t":48.07421875,"s":[0]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[183,56.999,0],"ix":2,"l":2},"a":{"a":0,"k":[16.667,3.333,0],"ix":1,"l":2},"s":{"a":1,"k":[{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":10.074,"s":[1005,1005,100]},{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":25.074,"s":[1235,1235,100]},{"t":48,"s":[0,0,100]}],"ix":6,"l":2}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0.102,0.897],[0.141,0.001],[0.016,-0.14],[0.44,-0.44],[0.912,-0.105],[0,-0.141],[-0.14,-0.016],[-0.451,-0.442],[-0.08,-0.906],[-0.144,0],[-0.013,0.144],[-0.451,0.451],[-0.89,0.077],[0,0.144],[0.143,0.013],[0.441,0.448]],"o":[[-0.016,-0.14],[-0.141,0],[-0.105,0.913],[-0.44,0.44],[-0.141,0.016],[0,0.141],[0.897,0.101],[0.448,0.441],[0.013,0.143],[0.144,0],[0.077,-0.891],[0.451,-0.451],[0.144,-0.013],[0,-0.144],[-0.906,-0.08],[-0.442,-0.451]],"v":[[0.276,-3.087],[0,-3.334],[-0.276,-3.088],[-1.094,-1.094],[-3.087,-0.276],[-3.333,0],[-3.087,0.276],[-1.084,1.092],[-0.277,3.08],[0,3.333],[0.277,3.079],[1.083,1.083],[3.079,0.277],[3.333,0],[3.08,-0.277],[1.092,-1.084]],"c":true},"ix":2},"nm":"路径 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.117647058824,0.588235294118,0.921568627451,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"填充 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[16.667,3.333],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"变换"}],"nm":"组 1","np":4,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":8400,"st":0,"ct":1,"bm":0}],"markers":[],"props":{}}'
-  background="transparent"
-  speed="1"
-  style="width: 300px; height: 300px;"
-  loop
-  autoplay
-></dotlottie-player>`;
-
 export const SmallHintIcon = html`<svg
   width="16"
   height="16"
@@ -361,3 +328,268 @@ export const ArrowUpIcon = html`<svg
     fill-opacity="0.6"
   />
 </svg> `;
+
+export const AIStarIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="none"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <g clip-path="url(#clip0_7444_26654)">
+    <path
+      d="M9.40033 4.57585C9.36627 4.27519 9.11207 4.04792 8.80948 4.04761C8.50689 4.0473 8.25223 4.27405 8.21756 4.57465C7.99211 6.52928 7.40771 7.90373 6.46492 8.84651C5.52213 9.7893 4.14769 10.3737 2.19305 10.5991C1.89246 10.6338 1.66571 10.8885 1.66602 11.1911C1.66633 11.4937 1.8936 11.7479 2.19426 11.7819C4.11659 11.9997 5.52025 12.5838 6.48542 13.5315C7.44696 14.4757 8.04356 15.8498 8.21597 17.7908C8.24327 18.0981 8.50096 18.3337 8.80954 18.3333C9.11813 18.333 9.37528 18.0969 9.40189 17.7894C9.56714 15.8804 10.1632 14.4774 11.1295 13.5111C12.0958 12.5448 13.4988 11.9487 15.4078 11.7835C15.7153 11.7569 15.9514 11.4997 15.9517 11.1911C15.9521 10.8825 15.7165 10.6249 15.4092 10.5976C13.4683 10.4252 12.0941 9.82855 11.15 8.86701C10.2022 7.90184 9.61807 6.49818 9.40033 4.57585Z"
+      fill="#1E96EB"
+    />
+    <path
+      d="M15.7849 1.87208C15.7717 1.75516 15.6728 1.66678 15.5551 1.66666C15.4375 1.66654 15.3384 1.75472 15.3249 1.87162C15.2373 2.63175 15.01 3.16626 14.6434 3.5329C14.2767 3.89954 13.7422 4.1268 12.9821 4.21448C12.8652 4.22796 12.777 4.327 12.7771 4.44467C12.7772 4.56234 12.8656 4.6612 12.9826 4.67445C13.7301 4.75912 14.276 4.9863 14.6513 5.35485C15.0253 5.72203 15.2573 6.25642 15.3243 7.01121C15.335 7.13075 15.4352 7.22235 15.5552 7.22221C15.6752 7.22208 15.7752 7.13025 15.7855 7.01069C15.8498 6.26831 16.0816 5.72268 16.4574 5.3469C16.8332 4.97111 17.3788 4.73932 18.1212 4.67505C18.2407 4.6647 18.3325 4.5647 18.3327 4.44469C18.3328 4.32469 18.2412 4.22448 18.1217 4.21386C17.3669 4.14681 16.8325 3.9148 16.4653 3.54087C16.0968 3.16552 15.8696 2.61966 15.7849 1.87208Z"
+      fill="#1E96EB"
+    />
+  </g>
+  <defs>
+    <clipPath id="clip0_7444_26654">
+      <rect width="20" height="20" fill="white" />
+    </clipPath>
+  </defs>
+</svg>`;
+
+export const AIStarIconWithAnimation = html`<dotlottie-player
+  src='{"v":"5.12.1","fr":120,"ip":0,"op":239,"w":240,"h":240,"nm":"ai","ddd":0,"assets":[],"layers":[{"ddd":0,"ind":1,"ty":4,"nm":"big 2","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":38,"s":[80]},{"t":76,"s":[100]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[107.859,132.143,0],"ix":2,"l":2},"a":{"a":0,"k":[8.572,11.429,0],"ix":1,"l":2},"s":{"a":1,"k":[{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":38,"s":[0,0,100]},{"t":76,"s":[1005,1005,100]}],"ix":6,"l":2}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0.261,2.307],[0.363,0],[0.041,-0.361],[1.131,-1.132],[2.346,-0.271],[-0.001,-0.363],[-0.361,-0.041],[-1.158,-1.137],[-0.207,-2.329],[-0.37,0],[-0.032,0.369],[-1.16,1.16],[-2.291,0.198],[-0.001,0.371],[0.368,0.033],[1.133,1.154]],"o":[[-0.041,-0.361],[-0.363,0],[-0.271,2.345],[-1.132,1.131],[-0.36,0.042],[0,0.363],[2.307,0.262],[1.154,1.133],[0.033,0.369],[0.37,0],[0.198,-2.29],[1.16,-1.159],[0.369,-0.032],[0,-0.37],[-2.329,-0.207],[-1.138,-1.158]],"v":[[0.71,-7.937],[0.001,-8.571],[-0.709,-7.938],[-2.812,-2.812],[-7.939,-0.709],[-8.571,0.001],[-7.937,0.71],[-2.788,2.81],[-0.711,7.921],[0.001,8.572],[0.712,7.919],[2.785,2.785],[7.919,0.712],[8.572,0.001],[7.921,-0.711],[2.81,-2.788]],"c":true},"ix":2},"nm":"路径 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.117647058824,0.588235294118,0.921568627451,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"填充 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[8.572,11.429],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"变换"}],"nm":"组 1","np":4,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":8400,"st":0,"ct":1,"bm":0},{"ddd":0,"ind":2,"ty":4,"nm":"big","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":0,"s":[100]},{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":15,"s":[80]},{"t":38,"s":[0]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[107.859,132.143,0],"ix":2,"l":2},"a":{"a":0,"k":[8.572,11.429,0],"ix":1,"l":2},"s":{"a":1,"k":[{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":0,"s":[1005,1005,100]},{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":15,"s":[1035,1035,100]},{"t":38,"s":[0,0,100]}],"ix":6,"l":2}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0.261,2.307],[0.363,0],[0.041,-0.361],[1.131,-1.132],[2.346,-0.271],[-0.001,-0.363],[-0.361,-0.041],[-1.158,-1.137],[-0.207,-2.329],[-0.37,0],[-0.032,0.369],[-1.16,1.16],[-2.291,0.198],[-0.001,0.371],[0.368,0.033],[1.133,1.154]],"o":[[-0.041,-0.361],[-0.363,0],[-0.271,2.345],[-1.132,1.131],[-0.36,0.042],[0,0.363],[2.307,0.262],[1.154,1.133],[0.033,0.369],[0.37,0],[0.198,-2.29],[1.16,-1.159],[0.369,-0.032],[0,-0.37],[-2.329,-0.207],[-1.138,-1.158]],"v":[[0.71,-7.937],[0.001,-8.571],[-0.709,-7.938],[-2.812,-2.812],[-7.939,-0.709],[-8.571,0.001],[-7.937,0.71],[-2.788,2.81],[-0.711,7.921],[0.001,8.572],[0.712,7.919],[2.785,2.785],[7.919,0.712],[8.572,0.001],[7.921,-0.711],[2.81,-2.788]],"c":true},"ix":2},"nm":"路径 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.117647058824,0.588235294118,0.921568627451,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"填充 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[8.572,11.429],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"变换"}],"nm":"组 1","np":4,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":8400,"st":0,"ct":1,"bm":0},{"ddd":0,"ind":3,"ty":4,"nm":"small 2","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":48.074,"s":[80]},{"t":85.07421875,"s":[100]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[183,56.999,0],"ix":2,"l":2},"a":{"a":0,"k":[16.667,3.333,0],"ix":1,"l":2},"s":{"a":1,"k":[{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":48,"s":[0,0,100]},{"t":85.07421875,"s":[1005,1005,100]}],"ix":6,"l":2}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0.102,0.897],[0.141,0.001],[0.016,-0.14],[0.44,-0.44],[0.912,-0.105],[0,-0.141],[-0.14,-0.016],[-0.451,-0.442],[-0.08,-0.906],[-0.144,0],[-0.013,0.144],[-0.451,0.451],[-0.89,0.077],[0,0.144],[0.143,0.013],[0.441,0.448]],"o":[[-0.016,-0.14],[-0.141,0],[-0.105,0.913],[-0.44,0.44],[-0.141,0.016],[0,0.141],[0.897,0.101],[0.448,0.441],[0.013,0.143],[0.144,0],[0.077,-0.891],[0.451,-0.451],[0.144,-0.013],[0,-0.144],[-0.906,-0.08],[-0.442,-0.451]],"v":[[0.276,-3.087],[0,-3.334],[-0.276,-3.088],[-1.094,-1.094],[-3.087,-0.276],[-3.333,0],[-3.087,0.276],[-1.084,1.092],[-0.277,3.08],[0,3.333],[0.277,3.079],[1.083,1.083],[3.079,0.277],[3.333,0],[3.08,-0.277],[1.092,-1.084]],"c":true},"ix":2},"nm":"路径 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.117647058824,0.588235294118,0.921568627451,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"填充 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[16.667,3.333],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"变换"}],"nm":"组 1","np":4,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":8400,"st":0,"ct":1,"bm":0},{"ddd":0,"ind":4,"ty":4,"nm":"small","sr":1,"ks":{"o":{"a":1,"k":[{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":10.074,"s":[100]},{"i":{"x":[0.5],"y":[1]},"o":{"x":[0.5],"y":[0]},"t":25.074,"s":[80]},{"t":48.07421875,"s":[0]}],"ix":11},"r":{"a":0,"k":0,"ix":10},"p":{"a":0,"k":[183,56.999,0],"ix":2,"l":2},"a":{"a":0,"k":[16.667,3.333,0],"ix":1,"l":2},"s":{"a":1,"k":[{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":10.074,"s":[1005,1005,100]},{"i":{"x":[0.5,0.5,0.5],"y":[1,1,1]},"o":{"x":[0.5,0.5,0.5],"y":[0,0,0]},"t":25.074,"s":[1235,1235,100]},{"t":48,"s":[0,0,100]}],"ix":6,"l":2}},"ao":0,"shapes":[{"ty":"gr","it":[{"ind":0,"ty":"sh","ix":1,"ks":{"a":0,"k":{"i":[[0.102,0.897],[0.141,0.001],[0.016,-0.14],[0.44,-0.44],[0.912,-0.105],[0,-0.141],[-0.14,-0.016],[-0.451,-0.442],[-0.08,-0.906],[-0.144,0],[-0.013,0.144],[-0.451,0.451],[-0.89,0.077],[0,0.144],[0.143,0.013],[0.441,0.448]],"o":[[-0.016,-0.14],[-0.141,0],[-0.105,0.913],[-0.44,0.44],[-0.141,0.016],[0,0.141],[0.897,0.101],[0.448,0.441],[0.013,0.143],[0.144,0],[0.077,-0.891],[0.451,-0.451],[0.144,-0.013],[0,-0.144],[-0.906,-0.08],[-0.442,-0.451]],"v":[[0.276,-3.087],[0,-3.334],[-0.276,-3.088],[-1.094,-1.094],[-3.087,-0.276],[-3.333,0],[-3.087,0.276],[-1.084,1.092],[-0.277,3.08],[0,3.333],[0.277,3.079],[1.083,1.083],[3.079,0.277],[3.333,0],[3.08,-0.277],[1.092,-1.084]],"c":true},"ix":2},"nm":"路径 1","mn":"ADBE Vector Shape - Group","hd":false},{"ty":"fl","c":{"a":0,"k":[0.117647058824,0.588235294118,0.921568627451,1],"ix":4},"o":{"a":0,"k":100,"ix":5},"r":1,"bm":0,"nm":"填充 1","mn":"ADBE Vector Graphic - Fill","hd":false},{"ty":"tr","p":{"a":0,"k":[16.667,3.333],"ix":2},"a":{"a":0,"k":[0,0],"ix":1},"s":{"a":0,"k":[100,100],"ix":3},"r":{"a":0,"k":0,"ix":6},"o":{"a":0,"k":100,"ix":7},"sk":{"a":0,"k":0,"ix":4},"sa":{"a":0,"k":0,"ix":5},"nm":"变换"}],"nm":"组 1","np":4,"cix":2,"bm":0,"ix":1,"mn":"ADBE Vector Group","hd":false}],"ip":0,"op":8400,"st":0,"ct":1,"bm":0}],"markers":[],"props":{}}'
+  background="transparent"
+  speed="1"
+  style="width: 300px; height: 300px;"
+  loop
+  autoplay
+></dotlottie-player>`;
+
+export const AIStopIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="none"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <g clip-path="url(#clip0_7496_34464)">
+    <path
+      fill-rule="evenodd"
+      clip-rule="evenodd"
+      d="M17.0827 10C17.0827 13.912 13.9114 17.0834 9.99935 17.0834C6.08733 17.0834 2.91602 13.912 2.91602 10C2.91602 6.088 6.08733 2.91669 9.99935 2.91669C13.9114 2.91669 17.0827 6.088 17.0827 10ZM18.3327 10C18.3327 14.6024 14.6017 18.3334 9.99935 18.3334C5.39698 18.3334 1.66602 14.6024 1.66602 10C1.66602 5.39765 5.39698 1.66669 9.99935 1.66669C14.6017 1.66669 18.3327 5.39765 18.3327 10ZM8.33268 6.66669C7.41221 6.66669 6.66602 7.41288 6.66602 8.33335V11.6667C6.66602 12.5872 7.41221 13.3334 8.33268 13.3334H11.666C12.5865 13.3334 13.3327 12.5872 13.3327 11.6667V8.33335C13.3327 7.41288 12.5865 6.66669 11.666 6.66669H8.33268Z"
+      fill="#1E96EB"
+    />
+  </g>
+  <defs>
+    <clipPath id="clip0_7496_34464">
+      <rect width="20" height="20" fill="white" />
+    </clipPath>
+  </defs>
+</svg>`;
+
+export const LanguageIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M7.50014 1.875C7.84532 1.875 8.12514 2.15482 8.12514 2.5V3.54167H10.6143C10.6216 3.54154 10.629 3.54154 10.6364 3.54167H12.5001C12.8453 3.54167 13.1251 3.82149 13.1251 4.16667C13.1251 4.51184 12.8453 4.79167 12.5001 4.79167H11.1458C10.6639 7.16518 9.64438 9.34282 8.22006 11.1922C8.5043 11.4886 8.80705 11.7614 9.12514 12C9.40128 12.2071 9.45724 12.5989 9.25014 12.875C9.04303 13.1511 8.65128 13.2071 8.37514 13C8.03723 12.7466 7.71615 12.4624 7.41408 12.1563C6.10818 13.5972 4.5372 14.7931 2.77828 15.6672C2.46917 15.8208 2.09406 15.6947 1.94044 15.3856C1.78683 15.0765 1.91288 14.7014 2.22199 14.5478C3.88561 13.721 5.36619 12.5816 6.58695 11.2065C5.71216 10.0729 5.07111 8.78956 4.74375 7.67633C4.64637 7.34517 4.83588 6.99777 5.16703 6.90039C5.49819 6.80301 5.84559 6.99252 5.94297 7.32367C6.2088 8.22765 6.71769 9.26651 7.3989 10.2054C8.57216 8.61889 9.42665 6.78272 9.86797 4.79167H2.9168C2.57163 4.79167 2.2918 4.51184 2.2918 4.16667C2.2918 3.82149 2.57163 3.54167 2.9168 3.54167H6.87514V2.5C6.87514 2.15482 7.15496 1.875 7.50014 1.875ZM13.3335 8.54167C13.5702 8.54167 13.7866 8.67542 13.8925 8.88716L16.8035 14.7093C16.8076 14.7169 16.8115 14.7247 16.8152 14.7326L18.0592 17.2205C18.2135 17.5292 18.0884 17.9046 17.7796 18.059C17.4709 18.2134 17.0955 18.0882 16.9411 17.7795L15.8639 15.625H10.8031L9.72582 17.7795C9.57145 18.0882 9.19603 18.2134 8.8873 18.059C8.57856 17.9046 8.45342 17.5292 8.60779 17.2205L9.85174 14.7326C9.85547 14.7247 9.85936 14.7169 9.86341 14.7093L12.7745 8.88716C12.8803 8.67542 13.0967 8.54167 13.3335 8.54167ZM11.4281 14.375H15.2389L13.3335 10.5642L11.4281 14.375Z"
+  />
+</svg> `;
+
+export const ImproveWritingIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M8.54167 2.5C8.54167 2.15482 8.26184 1.875 7.91667 1.875C7.57149 1.875 7.29167 2.15482 7.29167 2.5V5C7.29167 5.34518 7.57149 5.625 7.91667 5.625C8.26184 5.625 8.54167 5.34518 8.54167 5V2.5ZM8.54167 10.8333C8.54167 10.4882 8.26184 10.2083 7.91667 10.2083C7.57149 10.2083 7.29167 10.4882 7.29167 10.8333V13.3333C7.29167 13.6785 7.57149 13.9583 7.91667 13.9583C8.26184 13.9583 8.54167 13.6785 8.54167 13.3333V10.8333ZM5.625 7.91667C5.625 8.26184 5.34518 8.54167 5 8.54167H2.5C2.15482 8.54167 1.875 8.26184 1.875 7.91667C1.875 7.57149 2.15482 7.29167 2.5 7.29167H5C5.34518 7.29167 5.625 7.57149 5.625 7.91667ZM13.3333 8.54167C13.6785 8.54167 13.9583 8.26184 13.9583 7.91667C13.9583 7.57149 13.6785 7.29167 13.3333 7.29167H10.8333C10.4882 7.29167 10.2083 7.57149 10.2083 7.91667C10.2083 8.26184 10.4882 8.54167 10.8333 8.54167H13.3333ZM6.29623 6.29623C6.05215 6.54031 5.65642 6.54031 5.41235 6.29623L3.64458 4.52846C3.4005 4.28439 3.4005 3.88866 3.64458 3.64458C3.88866 3.4005 4.28439 3.4005 4.52846 3.64458L6.29623 5.41235C6.54031 5.65642 6.54031 6.05215 6.29623 6.29623ZM15.8081 16.6919C16.0521 16.936 16.4479 16.936 16.6919 16.6919C16.936 16.4479 16.936 16.0521 16.6919 15.8081L10.421 9.53712C10.1769 9.29304 9.7812 9.29304 9.53712 9.53712C9.29304 9.7812 9.29304 10.1769 9.53712 10.421L15.8081 16.6919ZM6.29623 9.5371C6.54031 9.78118 6.54031 10.1769 6.29623 10.421L4.52846 12.1888C4.28439 12.4328 3.88866 12.4328 3.64458 12.1888C3.4005 11.9447 3.4005 11.5489 3.64458 11.3049L5.41235 9.5371C5.65642 9.29303 6.05215 9.29303 6.29623 9.5371ZM12.1888 4.52845C12.4328 4.28437 12.4328 3.88864 12.1888 3.64457C11.9447 3.40049 11.549 3.40049 11.3049 3.64457L9.53712 5.41233C9.29304 5.65641 9.29304 6.05214 9.53712 6.29622C9.7812 6.54029 10.1769 6.54029 10.421 6.29622L12.1888 4.52845Z"
+  />
+</svg> `;
+
+export const AIDoneIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M16.2751 5.39143C16.5192 5.63551 16.5192 6.03124 16.2751 6.27532L7.94178 14.6086C7.6977 14.8527 7.30197 14.8527 7.0579 14.6086L3.72456 11.2753C3.48048 11.0312 3.48048 10.6355 3.72456 10.3914C3.96864 10.1474 4.36437 10.1474 4.60845 10.3914L7.49984 13.2828L15.3912 5.39143C15.6353 5.14735 16.031 5.14735 16.2751 5.39143Z"
+  />
+</svg> `;
+
+export const ShorterIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M4.1665 4.79163C3.82133 4.79163 3.5415 5.07145 3.5415 5.41663C3.5415 5.7618 3.82133 6.04163 4.1665 6.04163H15.8332C16.1783 6.04163 16.4582 5.7618 16.4582 5.41663C16.4582 5.07145 16.1783 4.79163 15.8332 4.79163H4.1665ZM4.1665 7.70829C3.82133 7.70829 3.5415 7.98811 3.5415 8.33329C3.5415 8.67847 3.82133 8.95829 4.1665 8.95829H9.99984C10.345 8.95829 10.6248 8.67847 10.6248 8.33329C10.6248 7.98811 10.345 7.70829 9.99984 7.70829H4.1665Z"
+  />
+</svg> `;
+
+export const LongerIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M4.1665 4.79163C3.82133 4.79163 3.5415 5.07145 3.5415 5.41663C3.5415 5.7618 3.82133 6.04163 4.1665 6.04163H15.8332C16.1783 6.04163 16.4582 5.7618 16.4582 5.41663C16.4582 5.07145 16.1783 4.79163 15.8332 4.79163H4.1665ZM4.1665 7.70829C3.82133 7.70829 3.5415 7.98811 3.5415 8.33329C3.5415 8.67847 3.82133 8.95829 4.1665 8.95829H15.8332C16.1783 8.95829 16.4582 8.67847 16.4582 8.33329C16.4582 7.98811 16.1783 7.70829 15.8332 7.70829H4.1665ZM3.5415 11.25C3.5415 10.9048 3.82133 10.625 4.1665 10.625H15.8332C16.1783 10.625 16.4582 10.9048 16.4582 11.25C16.4582 11.5951 16.1783 11.875 15.8332 11.875H4.1665C3.82133 11.875 3.5415 11.5951 3.5415 11.25ZM4.1665 13.5416C3.82133 13.5416 3.5415 13.8214 3.5415 14.1666C3.5415 14.5118 3.82133 14.7916 4.1665 14.7916H9.99984C10.345 14.7916 10.6248 14.5118 10.6248 14.1666C10.6248 13.8214 10.345 13.5416 9.99984 13.5416H4.1665Z"
+  />
+</svg> `;
+
+export const ToneIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M7.6411 7.08838L6.71487 9.4857C6.60938 9.75873 6.34683 9.93875 6.05413 9.93875H3.3335C2.98832 9.93875 2.7085 9.65893 2.7085 9.31375C2.7085 8.96857 2.98832 8.68875 3.3335 8.68875H5.68272L7.09896 5.02319C7.34911 4.37576 8.28457 4.43706 8.44809 5.11159L10.0805 11.8452L11.5785 7.1197C11.7512 6.57492 12.4648 6.44842 12.8142 6.90064L14.196 8.68875H16.6668C17.012 8.68875 17.2918 8.96857 17.2918 9.31375C17.2918 9.65893 17.012 9.93875 16.6668 9.93875H13.93C13.7106 9.93875 13.5036 9.83711 13.3695 9.66353L12.4575 8.48336L10.6861 14.0712C10.4695 14.7545 9.49134 14.7206 9.32248 14.024L7.6411 7.08838Z"
+  />
+</svg> `;
+
+export const ExplainIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M4.1665 4.375C3.82133 4.375 3.5415 4.65482 3.5415 5V15C3.5415 15.3452 3.82133 15.625 4.1665 15.625H5.83317C6.17835 15.625 6.45817 15.3452 6.45817 15C6.45817 14.6548 6.17835 14.375 5.83317 14.375H4.7915V5.625H5.83317C6.17835 5.625 6.45817 5.34518 6.45817 5C6.45817 4.65482 6.17835 4.375 5.83317 4.375H4.1665ZM15.8332 4.375C16.1783 4.375 16.4582 4.65482 16.4582 5V15C16.4582 15.3452 16.1783 15.625 15.8332 15.625H14.1665C13.8213 15.625 13.5415 15.3452 13.5415 15C13.5415 14.6548 13.8213 14.375 14.1665 14.375H15.2082V5.625H14.1665C13.8213 5.625 13.5415 5.34518 13.5415 5C13.5415 4.65482 13.8213 4.375 14.1665 4.375H15.8332Z"
+  />
+</svg> `;
+
+export const AIPenIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M16.0654 3.93417C15.5409 3.41078 14.6902 3.41078 14.1658 3.93417L13.4285 4.67003L15.3263 6.56414L16.0654 5.82654C16.5891 5.30384 16.5891 4.45687 16.0654 3.93417ZM14.4416 7.44716L12.5437 5.55304L3.5415 14.5375V16.4583H5.41262L14.4416 7.44716ZM13.2828 3.04941C14.2951 2.03903 15.936 2.03903 16.9484 3.04941C17.9614 4.06047 17.9614 5.70024 16.9484 6.7113L6.11265 17.5257C5.99547 17.6427 5.83669 17.7083 5.67114 17.7083H2.9165C2.57133 17.7083 2.2915 17.4285 2.2915 17.0833V14.2783C2.2915 14.1123 2.35752 13.9532 2.475 13.8359L13.2828 3.04941Z"
+  />
+</svg> `;
+
+export const ChatWithAIIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M3.12926 6.59386C2.2839 5.82997 2.79173 4.375 3.97096 4.375H15.8351C16.8185 4.375 17.3998 5.44049 16.9328 6.27292L11.1139 16.6458C10.543 17.6636 9.04423 17.4269 8.78738 16.306L7.45823 10.5057L3.12926 6.59386ZM8.73556 10.4835L10.0058 16.0268C10.0075 16.0342 10.0092 16.0381 10.01 16.0396C10.011 16.0401 10.013 16.041 10.0161 16.0414C10.0172 16.0416 10.0182 16.0417 10.0191 16.0417C10.0202 16.0401 10.0218 16.0377 10.0238 16.0343L15.26 6.70001L8.73556 10.4835ZM14.622 5.625H3.97096C3.96952 5.625 3.96831 5.62504 3.96732 5.6251C3.96581 5.62713 3.96365 5.63059 3.96169 5.63585C3.95772 5.64646 3.95838 5.65425 3.95884 5.65639C3.95897 5.65699 3.95903 5.65711 3.95919 5.65738L3.9592 5.6574C3.95926 5.6575 3.961 5.66071 3.96732 5.66642L8.10424 9.40467L14.622 5.625Z"
+  />
+</svg> `;
+
+export const AISearchIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M3.125 7.91667C3.125 5.2703 5.2703 3.125 7.91667 3.125C10.563 3.125 12.7083 5.2703 12.7083 7.91667C12.7083 10.563 10.563 12.7083 7.91667 12.7083C5.2703 12.7083 3.125 10.563 3.125 7.91667ZM7.91667 1.875C4.57995 1.875 1.875 4.57995 1.875 7.91667C1.875 11.2534 4.57995 13.9583 7.91667 13.9583C9.35991 13.9583 10.685 13.4523 11.724 12.6079L16.2247 17.1086C16.4688 17.3527 16.8645 17.3527 17.1086 17.1086C17.3527 16.8645 17.3527 16.4688 17.1086 16.2247L12.6079 11.724C13.4523 10.685 13.9583 9.35991 13.9583 7.91667C13.9583 4.57995 11.2534 1.875 7.91667 1.875Z"
+  />
+</svg> `;
+
+export const MakeItRealIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M1.89506 9.08909C1.36098 9.73568 1.85634 10.7079 2.69338 10.6559L5.70784 10.4685L7.3281 13.0174C7.77801 13.7251 8.85573 13.5545 9.06491 12.7423L9.63712 10.5207L15.8083 16.6918C16.0524 16.9359 16.4481 16.9359 16.6922 16.6918C16.9362 16.4477 16.9362 16.052 16.6922 15.8079L10.5208 9.63654L12.743 9.06417C13.5552 8.855 13.7259 7.77728 13.0181 7.32737L10.4692 5.7071L10.6566 2.69265C10.7086 1.85562 9.73642 1.36024 9.08982 1.89433L6.76121 3.81777L3.9522 2.70805C3.17221 2.39991 2.40064 3.17147 2.70878 3.95147L3.8185 6.76048L1.89506 9.08909ZM5.80215 9.21024L3.28715 9.36656L4.8919 7.42377C5.11419 7.15466 5.17258 6.78598 5.04433 6.46135L4.11848 4.11775L6.46208 5.0436C6.78671 5.17185 7.15539 5.11346 7.4245 4.89117L9.36729 3.28642L9.21097 5.80142C9.18931 6.14979 9.35878 6.48238 9.65335 6.66963L11.7799 8.02144L9.33969 8.64995C9.00168 8.737 8.73774 9.00095 8.65068 9.33896L8.02217 11.7792L6.67036 9.65261C6.48312 9.35805 6.15052 9.18858 5.80215 9.21024Z"
+  />
+</svg> `;
+
+export const TagIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M3.80314 6.22236C3.85186 4.90697 4.90697 3.85186 6.22236 3.80314L10.0611 3.66097C10.3298 3.65102 10.5905 3.75339 10.7807 3.94351L16.3203 9.48317C16.6979 9.86077 16.6979 10.473 16.3203 10.8506L10.8506 16.3203C10.473 16.6979 9.86076 16.6979 9.48316 16.3203L3.94351 10.7807C3.75339 10.5905 3.65101 10.3298 3.66097 10.0612L3.80314 6.22236ZM6.17942 2.64364C4.25692 2.71484 2.71484 4.25693 2.64364 6.17942L2.50146 10.0182C2.47957 10.6093 2.70479 11.1829 3.12305 11.6011L8.6627 17.1408C9.49343 17.9715 10.8403 17.9715 11.671 17.1408L17.1408 11.6711C17.9715 10.8403 17.9715 9.49344 17.1408 8.66271L11.6011 3.12305C11.1828 2.70479 10.6093 2.47957 10.0182 2.50146L6.17942 2.64364ZM6.88504 6.88505C7.18713 6.58296 7.18713 6.09319 6.88504 5.7911C6.58296 5.48902 6.09319 5.48902 5.7911 5.7911C5.48902 6.09319 5.48902 6.58296 5.7911 6.88505C6.09318 7.18713 6.58296 7.18713 6.88504 6.88505Z"
+  />
+</svg> `;
+
+export const EnterIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M11.2502 4.16663C11.2502 3.82145 11.53 3.54163 11.8752 3.54163H15.6252C15.9703 3.54163 16.2502 3.82145 16.2502 4.16663V11.25C16.2502 11.4157 16.1843 11.5747 16.0671 11.6919C15.9499 11.8091 15.7909 11.875 15.6252 11.875L5.45567 11.875L8.24903 14.6994C8.49175 14.9448 8.48957 15.3405 8.24414 15.5832C7.99872 15.826 7.603 15.8238 7.36027 15.5783L3.51412 11.6895C3.27329 11.4459 3.27329 11.054 3.51412 10.8105L7.36027 6.92158C7.603 6.67615 7.99872 6.67397 8.24414 6.91669C8.48957 7.15942 8.49175 7.55514 8.24903 7.80056L5.45567 10.625L15.0002 10.625V4.79163H11.8752C11.53 4.79163 11.2502 4.5118 11.2502 4.16663Z"
+  />
+</svg> `;
+
+export const ArrowRightIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M7.05838 5.39143C6.81431 5.63551 6.81431 6.03124 7.05838 6.27532L10.7831 10L7.05838 13.7248C6.81431 13.9688 6.81431 14.3646 7.05838 14.6087C7.30246 14.8527 7.69819 14.8527 7.94227 14.6087L12.1089 10.442C12.353 10.1979 12.353 9.80218 12.1089 9.5581L7.94227 5.39143C7.69819 5.14735 7.30246 5.14735 7.05838 5.39143Z"
+  />
+</svg> `;
+
+export const DiscardIcon = html`<svg
+  width="20"
+  height="20"
+  viewBox="0 0 20 20"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M9.16667 1.87512C7.90101 1.87512 6.875 2.90114 6.875 4.16679V5.12504H3.33333C2.94213 5.12504 2.625 5.44217 2.625 5.83337C2.625 6.22458 2.94213 6.54171 3.33333 6.54171H3.59066L4.26602 15.9967C4.35168 17.196 5.34957 18.1251 6.55186 18.1251H13.4481C14.6504 18.1251 15.6483 17.196 15.734 15.9967L16.4093 6.54171H16.6667C17.0579 6.54171 17.375 6.22458 17.375 5.83337C17.375 5.44217 17.0579 5.12504 16.6667 5.12504H13.125V4.16679C13.125 2.90114 12.099 1.87512 10.8333 1.87512H9.16667ZM11.875 5.12504V4.16679C11.875 3.59149 11.4086 3.12512 10.8333 3.12512H9.16667C8.59137 3.12512 8.125 3.59149 8.125 4.16679V5.12504H11.875ZM15.1562 6.54171H4.84385L5.51285 15.9077C5.55178 16.4528 6.00537 16.8751 6.55186 16.8751H13.4481C13.9946 16.8751 14.4482 16.4528 14.4872 15.9077L15.1562 6.54171ZM8.95833 9.16679C8.95833 8.82161 8.67851 8.54179 8.33333 8.54179C7.98815 8.54179 7.70833 8.82161 7.70833 9.16679V14.1668C7.70833 14.512 7.98815 14.7918 8.33333 14.7918C8.67851 14.7918 8.95833 14.512 8.95833 14.1668V9.16679ZM11.6667 8.54179C12.0118 8.54179 12.2917 8.82161 12.2917 9.16679V14.1668C12.2917 14.512 12.0118 14.7918 11.6667 14.7918C11.3215 14.7918 11.0417 14.512 11.0417 14.1668V9.16679C11.0417 8.82161 11.3215 8.54179 11.6667 8.54179Z"
+  />
+</svg>`;
