+++ packages/presets/src/ai/entries/edgeless/panel-config.ts
@@ -1,7 +1,8 @@
-import type { AffineAIPanelWidgetConfig } from '@blocksuite/blocks';
-import type { AffineAIPanelWidget } from '@blocksuite/blocks';
-
-import { createTextRenderer } from '../../messages/text.js';
+import type {
+  AffineAIPanelWidget,
+  AffineAIPanelWidgetConfig,
+} from '../../widgets/ai-panel/ai-panel.js';
+import { createTextRenderer } from '../../widgets/messages/text.js';
 
 export function buildEdgelessPanelConfig(
   panel: AffineAIPanelWidget
