+++ packages/presets/src/ai/index.ts
@@ -1,7 +1,3 @@
 export * from './ai.js';
-export { ChatPanel } from './chat-panel/index.js';
-export * from './entries/edgeless/actions-config.js';
-export * from './entries/index.js';
-export * from './messages/index.js';
 export * from './provider.js';
 export * from './utils/event-source.js';
