+++ packages/presets/src/ai/widgets/ai-item/styles.ts
