+++ packages/presets/src/ai/widgets/ai-item/types.ts
@@ -1,7 +1,6 @@
 import type { Chain, EditorHost, InitCommandCtx } from '@blocksuite/block-std';
+import type { EditorMode } from '@blocksuite/blocks';
 import type { TemplateResult } from 'lit';
-
-import type { EditorMode } from '../../utils/index.js';
 
 export interface AIItemGroupConfig {
   name?: string;
