+++ packages/presets/src/ai/widgets/ai-panel/ai-panel.ts
