+++ packages/presets/src/ai/widgets/ai-panel/components/divider.ts
