+++ packages/presets/src/ai/widgets/ai-panel/components/state/index.ts
